+++ sdk/cosmos/azure_data_cosmos/src/clients/cosmos_client.rs
@@ -146,7 +146,7 @@ impl CosmosClient {
         let plan = self
             .context
             .driver
-            .plan_operation(initial_operation, &operation_options, None)
+            .plan_operation(initial_operation, &operation_options, None, None)
             .await?;
 
         Ok(QueryItemIterator::new(