+++ sdk/cosmos/azure_data_cosmos/src/clients/database_client.rs
@@ -131,7 +131,7 @@ impl DatabaseClient {
         let plan = self
             .context
             .driver
-            .plan_operation(initial_operation, &operation_options, None)
+            .plan_operation(initial_operation, &operation_options, None, None)
             .await?;
 
         Ok(QueryItemIterator::new(